prototype needs to be in main file
